Stop Trace from clipping the first function in the stacktrace
Trace will trim the stack trace it prints so that you don't always see the Belt code as part of the stack trace. Something changed, probably PHP 7 that made it so there was one fewer function call in the trace, so now it's clipping the first part of the real stacktrace.

Adjusted the defaults so that it's just _generateLine that's setting the number of functions to trim.